# swagger/api.yml
openapi: 3.0.3
servers:
  - description: API Localhost Debug
    url: http://localhost:8080
  - description: API Production talfekler.com
    url: https://talfekler.com
info:
  description: Tal Studio API
  version: 1.0.0
  title: Tal Studio
  contact:
    email: berdugogo@gmail.com
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - "auth"
      summary: "User login into system"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/user_login"
            examples:
              user_berdugogo:
                $ref: "#/components/examples/user_login_berdugogo"
      responses:
        "200":
          description: "התחברות התבצעה בהצלחה"
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: session=8dsf8esaf65c1xdf3ax2asdcz3asd
        "401":
          description: "פרטי התחברות שגויים"
  /api/auth/logout:
    post:
      security:
        - cookieAuth: []
      tags:
        - "auth"
      summary: "logout user"
      description: "logging out a user and deletes his cookie."
      responses:
        "200":
          description: "משתמש התנתק בהצלחה"
        "401":
          description: "בעיה בחיבור לחשבונך. יש להתחבר שוב. (ייתכן כיהבעיה נגרמת מכך שהמשתמש איננו מחובר כבר בעת שניסה להתנתק)"
  /api/auth/register:
    post:
      tags:
        - "auth"
      summary: "Create user"
      description: "Added a user into the database"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/reg_user"
            examples:
              user_berdugogo:
                $ref: "#/components/examples/user_full_berdugogo_reg"
      responses:
        "201":
          description: "המשתמש נוצר בהצלחה. מייל נשלח."
        "409":
          description: "המייל שנבחר בשימוש.| הטלפון שנבחר בשימוש. יש לבחור אחר"
  /api/auth/validatepass:
    post:
      tags:
        - "auth"
      security:
        - cookieAuth: []
      summary: "Validate password only, to allow Profile edit"
      requestBody:
        description: "the user password entered. without client check"
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                password:
                  type: string
                  example: 123123
      responses:
        "200":
          description: "סיסמה תקינה"
        "402":
          description: "סיסמה שגויה"
        "401":
          description: "בעיה בחיבור לחשבונך. יש להתחבר שוב"
  /api/auth/logged:
    get:
      security:
        - cookieAuth: []
      tags:
        - "auth"
      summary: "full user details on log"
      responses:
        "200":
          description: "user_full by cookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/reg_user"
              examples:
                berdugogo@gmail.com:
                  $ref: "#/components/examples/user_full_berdugogo"
        "401":
          description: "בעיה בחיבור לחשבונך. יש להתחבר שוב"
  /api/auth/updateprofile:
    put:
      security:
        - cookieAuth: []
      tags:
        - "auth"
      summary: "Update Profile Request by authenticated user."
      requestBody:
        description: JSON object to update. Same rules of register.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/reg_user"
            examples:
              user_berdugogo:
                $ref: "#/components/examples/user_full_berdugogo_reg"
      responses:
        "200":
          description: "updated user_full"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/reg_user"
              examples:
                berdugogo@gmail.com:
                  $ref: "#/components/examples/user_full_berdugogo"
        "400":
          description: "אחד הפרטים שהוזן אינו תקין"
        "401":
          description: "בעיה בחיבור לחשבונך. יש להתחבר שוב"
        "409":
          description: "המייל|הטלפון שנבחר בשימוש. יש לבחור אחר"
        "501":
          description: "שגיאה בבסיס הנתונים, אנא נסו שנית מאוחר יותר"
        "500":
          description: "שגיאה בשרת, אנא נסו שנית מאוחר יותרר"
...
  /api/admin/available/add:
    post:
      tags:
        - "admin"
      security:
        - cookieAuth: []
      summary: "admin add new open trainings"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                events:
                  type: array
                  items:
                    type: object
                    properties:
                      date: 
                        type: string
                        format: DATETIME
              example: 
                {
                  events: 
                    [
                      date: "2021-10-31T04:00:00.000Z",
                      date: "2021-10-31T05:00:00.000Z"
                    ]
                }
      responses:
        "201":
          description: "האימונים הפנויים נשמרו בהצלחה"
        "400":
          description: "Cannot retrieve trainings"
        "403":
          description: "שמירת אימונים פנויים נכשלה"
  /api/admin/next_events_dates:
    get:
      tags:
        - "admin"
      summary: "get all next events and trainings globally splited by dates"
      responses:
        "201":
          content:
            application/json:
              examples:
                next_events_globally_ex1:
                  $ref: "#/components/examples/next_events_dates"
          description: "Trainings retrieved successfully"
        "400":
          description: "Cannot retrieve trainings"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: session
      in: cookie
  schemas:
...
    reg_user:
      type: "object"
      properties:
        city:
          type: string
        email:
          type: string
          format: email
        fav_sport:
          type: string
        firstname:
          type: string
        gender:
          type: string
        lastname:
          type: string
        password:
          type: string
          minLength: 6
          maxLength: 12
        phone:
          type: string
          minLength: 10
          maxLength: 10
        sport_for_me:
          type: string
        training_num:
          type: integer
          minimum: 0
          maximum: 7
        zip:
          type: integer
          maxLength: 10
        birthdate:
          type: string
          format: string
    user_login:
      type: "object"
      properties:
        email:
          type: string
          format: email
        password:
          description: "The password for login in clear text,5-10 chars, at least one number and one special character"
          type: string
          format: password
          minLength: 6
          maxLength: 12
...
  examples:
    user_full_berdugogo:
      value:
        role: "trainer"
        _id: "6173232477f5772770ccc9c9"
        city: "תל אביב - יפו"
        firstname: "עדן"
        lastname: "ברדוגו"
        email: "edenjunk94@gmail.com"
        fav_sport: "functional"
        gender: "female"
        phone: "0547585528"
        sport_for_me: "hobby"
        training_num: "5"
        zip: "3436643"
        birthdate: "24/10/1994"
    user_full_berdugogo_reg:
        value:
          city: "תל אביב - יפו"
          firstname: "עדן"
          lastname: "ברדוגו"
          email: "edenjunk94@gmail.com"
          fav_sport: "functional"
          gender: "female"
          phone: "0547585528"
          sport_for_me: "hobby"
          training_num: 5
          zip: "3436643"
          birthdate: "24/10/1994"
          password: "123123e"
    user_login_berdugogo:
      value:
        email: "berdugogo@gmail.com"
        password: "123123e"
...
    next_events_dates:
      value:
        {
          "2021-10-24": 
            [
              {
                _id: "61734a7dfa6df322148fea9c",
                date: "2021-10-24T06:00:00.000Z",
                title: "פנוי",
                description: ""
              },
              {
                _id: "61734a7dfa6df322148fea9d",
                date: "2021-10-24T07:00:00.000Z",
                title: "פנוי",
                description: ""
              },        
            ],
          "2021-10-29": [
            {
              _id: "6173477c70fc735c5004d009",
              date: "2021-10-29T03:00:00.000Z",
              title: "פנוי",
              description: ""
            }
          ],
        }
...